backend/api/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import UserViewSet, TransactionViewSet, GlobalStatsViewSet

router = DefaultRouter()
router.register(r'users', UserViewSet, basename='user')
router.register(r'transactions', TransactionViewSet, basename='transaction')
router.register(r'stats', GlobalStatsViewSet, basename='stats')

urlpatterns = [
    path('', include(router.urls)),
]





















