backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Production Server
gunicorn==21.2.0

# Static Files
whitenoise==6.6.0

# Image Processing
Pillow>=10.3.0

# Environment Variables
python-decouple==3.8

# Authentication
google-auth==2.23.4
requests==2.31.0
